+++ changelog.md
@@ -9,6 +9,7 @@
   - Added Deep Fungal Cobblestone sets
   - Added Hard Deep Fungal Block sets
   - Added Rocky Deep Fungal Stone sets
+- Added Picket Fences 
 
 Please report any bugs found to [GitHub](https://github.com/WenXin20/Marioverse/issues)